requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
groq==0.4.1
PyPDF2==3.0.1
requests==2.31.0
gTTS==2.4.0
